Fix some object event load functions that don't take the expansion into account and implement dynamic mugshot battle transition

+++ src/apprentice.c
@@ -1179,7 +1179,7 @@ static void SaveApprentice(void)
 static void SetPlayerApprenticeTrainerGfxId(void)
 {
     u32 i;
-    u8 objectEventGfxId;
+    u16 objectEventGfxId;
     u8 class = gApprentices[PLAYER_APPRENTICE.id].facilityClass;
 
     for (i = 0; i < ARRAY_COUNT(gTowerMaleFacilityClasses) && gTowerMaleFacilityClasses[i] != class; i++)
